feat(spec-renderer): move spec-operations-list package
