--- packages/portal/spec-operations-list/package.json
@@ -1,60 +1,0 @@
-{
-  "name": "@kong-ui-public/spec-operations-list",
-  "version": "0.6.4",
-  "type": "module",
-  "main": "./dist/spec-operations-list.umd.js",
-  "module": "./dist/spec-operations-list.es.js",
-  "types": "dist/types/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "exports": {
-    ".": {
-      "import": "./dist/spec-operations-list.es.js",
-      "require": "./dist/spec-operations-list.umd.js"
-    },
-    "./package.json": "./package.json",
-    "./dist/*": "./dist/*"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "scripts": {
-    "dev": "cross-env USE_SANDBOX=true vite",
-    "build": "run-s typecheck build:package build:types",
-    "build:package": "vite build",
-    "build:types": "vue-tsc -p './tsconfig.build.json' --emitDeclarationOnly",
-    "build:visualize": "BUILD_VISUALIZER='portal/spec-operations-list' vite build -m production",
-    "preview:package": "vite preview --port 4173",
-    "preview": "cross-env USE_SANDBOX=true PREVIEW_SANDBOX=true run-s build:package preview:package",
-    "lint": "eslint '**/*.{js,jsx,ts,tsx,vue}' --ignore-path '../../../.eslintignore'",
-    "lint:fix": "eslint '**/*.{js,jsx,ts,tsx,vue}' --ignore-path '../../../.eslintignore' --fix",
-    "stylelint": "stylelint --allow-empty-input './src/**/*.{css,scss,sass,less,styl,vue}'",
-    "stylelint:fix": "stylelint --allow-empty-input './src/**/*.{css,scss,sass,less,styl,vue}' --fix",
-    "typecheck": "vue-tsc -p './tsconfig.build.json' --noEmit",
-    "test:component": "BABEL_ENV=cypress cross-env FORCE_COLOR=1 cypress run --component -b chrome --spec './src/**/*.cy.ts' --project '../../../.'",
-    "test:component:open": "BABEL_ENV=cypress cross-env FORCE_COLOR=1 cypress open --component -b chrome --project '../../../.'"
-  },
-  "peerDependencies": {
-    "@kong/kongponents": "^8.22.2",
-    "vue": "^3.2.47"
-  },
-  "devDependencies": {
-    "@kong/kongponents": "^8.22.2",
-    "vue": "^3.2.47"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Kong/public-ui-components.git",
-    "directory": "packages/portal/spec-operations-list"
-  },
-  "homepage": "https://github.com/Kong/public-ui-components/tree/main/packages/portal/spec-operations-list",
-  "bugs": {
-    "url": "https://github.com/Kong/public-ui-components/issues"
-  },
-  "author": "Kong, Inc.",
-  "license": "Apache-2.0",
-  "volta": {
-    "extends": "../../../package.json"
-  }
-}
