--- packages/portal/spec-operations-list/src/global-components.d.ts
@@ -1,2 +1,0 @@
-// Import globally available components
-import '@kong/kongponents/dist/types/global-components'
