--- packages/portal/spec-operations-list/src/index.ts
@@ -1,5 +1,0 @@
-import OperationsList from './components/OperationsList.vue'
-
-export default OperationsList
-
-export * from './types'
