--- packages/portal/spec-operations-list/tsconfig.build.json
@@ -1,12 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "types": []
-  },
-  "exclude": [
-    "src/**/*.cy.ts",
-    "src/**/*.spec.ts",
-    "sandbox",
-    "dist"
-  ]
-}
