+++ packages/portal/spec-renderer/package.json
@@ -38,15 +38,16 @@
     "test:unit:open": "cross-env FORCE_COLOR=1 vitest --ui"
   },
   "peerDependencies": {
+    "@kong/kongponents": "^8.22.0",
     "vue": "^3.2.45"
   },
   "devDependencies": {
+    "@kong/kongponents": "^8.22.0",
     "@modyfi/vite-plugin-yaml": "^1.0.4",
     "openapi-types": "^12.1.0",
     "vue": "^3.2.45"
   },
   "dependencies": {
-    "@kong-ui-public/spec-operations-list": "^0.6.1",
     "@kong-ui-public/swagger-ui-web-component": "workspace:^0.3.0"
   },
   "repository": {
