+++ packages/portal/spec-renderer/src/index.ts
@@ -1,9 +1,11 @@
 import SpecRenderer from './components/SpecRenderer.vue'
 import SpecDetails from './components/SpecDetails.vue'
+import SpecOperationsList from './components/SpecOperationsList.vue'
 
 export {
   SpecRenderer,
   SpecDetails,
+  SpecOperationsList,
 }
 
 export * from './types'
