+++ packages/portal/spec-renderer/src/types/spec-renderer.ts
@@ -10,14 +10,36 @@ export type SpecDocument = OpenAPI.Document
  */
 export type MethodString = 'get' | 'put' | 'post' | 'delete' | 'patch' | 'options' | 'head' | 'connect' | 'trace'
 
-/**
- * Operation details object.
- */
 export interface Operation {
   path: string
   method: MethodString
   operationId: string | null
   summary: string | null
   deprecated: boolean
+  tags: string[]
+}
+
+/**
+ * Operation details object.
+ */
+export interface OperationListItem extends Omit<Operation, 'tags'> {
   tag: string | null
 }
+
+interface ExternalDocumentation {
+  description?: string
+  url: string
+}
+
+export interface Tag {
+  name: string
+  description?: string
+  externalDocs?: ExternalDocumentation
+}
+
+export interface OperationListFilterFunctionParams {
+  items: Operation[]
+  query: string
+}
+
+export type OperationListFilterFunction = (params: OperationListFilterFunctionParams) => Operation[]
